+++ tests/mocks.py
@@ -74,14 +74,19 @@ SAT_VERSIONS = {
 
 # Two versions of a product named COS where:
 # - The two versions have one docker image name and version in common
-# - The first version has no repositories, configuration, images, or recipes
-# - The second version has repositories, configuration, images, and recipes
+# - The first version has docker images and manifests but not helm charts, repositories, configuration,
+#   images, or recipes
+# - The second version has docker images, helm charts, repositories, configuration, images, and recipes,
+#   but not manifests
 COS_VERSIONS = {
     '2.0.0': {
         'component_versions': {
             'docker': [
                 {'name': 'cray/cray-cos', 'version': '1.0.0'},
                 {'name': 'cray/cos-cfs-install', 'version': '1.4.0'}
+            ],
+            'manifests': [
+                'config-data/argo/loftsman/cos/2.0.0/manifests/cos-services.yaml'
             ]
         }
     },
@@ -90,6 +95,11 @@ COS_VERSIONS = {
             'docker': [
                 {'name': 'cray/cray-cos', 'version': '1.0.1'},
                 {'name': 'cray/cos-cfs-install', 'version': '1.4.0'}
+            ],
+            'helm': [
+                {'name': 'cos-config', 'version': '0.4.76'},
+                {'name': 'cos-sle15sp3-artifacts', 'version': '1.3.23'},
+                {'name': 'cray-cps', 'version': '1.8.15'}
             ],
             'repositories': [
                 {'name': 'cos-sle-15sp2', 'type': 'group', 'members': ['cos-2.0.1-sle-15sp2']},
@@ -116,6 +126,39 @@ COS_VERSIONS = {
     },
 }
 
+# One version of "cpe" product that has repositories, s3, configuration, images, and recipes
+CPE_VERSION = {
+    '2.0.0': {
+        'component_versions': {
+            'repositories': [
+                {'name': 'cpe-2.0-sles15-sp4', 'type': 'hosted'},
+                {'name': 'cpe-sles15-sp4', 'type': 'group', 'members': ['cpe-2.0-sles15-sp4']}
+            ],
+            's3': [
+                {'bucket': 'boot-images', 'key': 'PE/CPE-base.x86_64-2.0.squashfs'},
+                {'bucket': 'boot-images', 'key': 'PE/CPE-amd.x86_64-2.0.squashfs'}
+            ]
+        },
+        "configuration": {
+            "clone_url": "https://vcs.cmn.lemondrop.hpc.amslabs.hpecorp.net/vcs/cray/cpe-config-management.git",
+            "commit": "cbfa9669ed253499406f268022b0081509036906",
+            "import_branch": "cray/cpe/2.0.0",
+            "import_date": "2023-03-23T12:04:25.210495Z",
+            "ssh_url": "git@vcs.machine.dev.cray.com:cray/cpe-config-management.git"
+        },
+        "images": {
+            "cpe-barebones-sles15sp4.x86_64-2.0.0": {
+                "id": "cf9c87ea-014c-448a-86ac-3ef3e9d2178f"
+            }
+        },
+        "recipes": {
+            "cpe-barebones-sles15sp4.x86_64-2.0.0": {
+                "id": "a30ad9b4-a1ce-4734-bbc1-6de2c1e3781a"
+            }
+        }
+    }
+}
+
 # One version of "Other Product" that also uses cray/cray-sat:1.0.1
 OTHER_PRODUCT_VERSION = {
     '2.0.0': {
@@ -131,10 +174,38 @@ OTHER_PRODUCT_VERSION = {
     }
 }
 
+# Multiple versions of products named 'cos', 'sat', and 'cpe' that has valid YAML but not matching schema
+# - 'cos' product with invalid component manifests
+# - 'sat' product with invalid component docker
+# - 'cpe' product with invalid component s3
+MOCK_INVALID_PRODUCT_DATA = {
+    'cos': {
+        '2.1': {
+            'component_versions': {
+                'manifests': 'should be an array'
+            }
+        }
+    },
+    'sat': {
+        '2.1': {
+            'component_versions': {
+                'docker': 'should be an array'
+            }
+        }
+    },
+    'cpe': {
+        '2.1': {
+            'component_versions': {
+                's3': 'should be an array'
+            }
+        }
+    }
+}
 
 # A mock version of the data returned when querying the Product Catalog ConfigMap
 MOCK_PRODUCT_CATALOG_DATA = {
     'sat': safe_dump(SAT_VERSIONS),
     'cos': safe_dump(COS_VERSIONS),
+    'cpe': safe_dump(CPE_VERSION),
     'other_product': safe_dump(OTHER_PRODUCT_VERSION)
 }